+++ src/flag_gems/ops/ge.py
@@ -8,7 +8,7 @@ from ..utils import unwrap
 
 @triton.jit
 def ge_func(x, y):
-    return x.to(tl.float32) >= y
+    return x >= y
 
 def ge(A, B):
     logging.debug("GEMS GE")
