+++ src/flag_gems/ops/gt.py
@@ -8,7 +8,7 @@ from ..utils import unwrap
 
 @triton.jit
 def gt_func(x, y):
-    return x.to(tl.float32) > y
+    return x > y
 
 def gt(A, B):
     logging.debug("GEMS GT")
